+++ allium/src/main/java/dev/hugeblank/allium/loader/mixin/MixinClassInfo.java
@@ -4,6 +4,7 @@ import dev.hugeblank.allium.AlliumPreLaunch;
 import dev.hugeblank.allium.loader.type.StaticBinder;
 import dev.hugeblank.allium.loader.type.annotation.LuaWrapped;
 import dev.hugeblank.allium.util.Identifiable;
+import dev.hugeblank.allium.util.MixinConfigUtil;
 import me.basiqueevangelist.enhancedreflection.api.EClass;
 import org.squiddev.cobalt.LuaValue;
 
@@ -29,7 +30,7 @@ public class MixinClassInfo implements Identifiable {
 
     @LuaWrapped
     public LuaValue getInterface() throws ClassNotFoundException {
-        if (!AlliumPreLaunch.isComplete()) throw new IllegalStateException("Mixin cannot be accessed in pre-launch phase.");
+        if (!MixinConfigUtil.isComplete()) throw new IllegalStateException("Mixin cannot be accessed in pre-launch phase.");
         if (!isInterface) throw new IllegalStateException("Cannot get interface of non-interface mixin.");
         EClass<?> clazz = EClass.fromJava(Class.forName(className));
         return StaticBinder.bindClass(clazz);
