+++ allium/src/main/java/dev/hugeblank/allium/util/asm/VisitedField.java
@@ -1,3 +1,9 @@
 package dev.hugeblank.allium.util.asm;
 
-public record VisitedField(VisitedClass owner, int access, String name, String descriptor, String signature, Object value) implements VisitedElement { }
+import static org.objectweb.asm.Opcodes.ACC_STATIC;
+
+public record VisitedField(VisitedClass owner, int access, String name, String descriptor, String signature, Object value) implements VisitedElement {
+    public boolean isStatic() {
+        return (access() & ACC_STATIC) != 0;
+    }
+}
